refactor web table logic into reusable modules
